Update update.sh
Added authentication functionality so that premium users can take advantage on premium support and updates. Also increases security as bots and hackers will not use the code that is only allowed for commercial usage for illegal purposes. Saved bandwidth on hosting servers and dedicated it for only other users that has subscribed to our services and update their machines as soon as possible.