Update update.sh
Many changes and now it tells the user verbosely about the processes being executed. It can be used to pipe to other development files and other programs. The output can serve as a helpful API for other software. Added cleaning up functionality that will remove all unwanted packages and now it gets a list of all updates, rather than just one update in the older release.